validate ratechecker loader using external file
This commit modifies how the `load_daily_data` management command
validates the new data that it loads. Instead of using a set of
hardcoded scenarios defined in a Python module, it instead reads in an
external file defining the scenarios to use.

The external file must be in JSONL format (file where each line is JSON
formatted) and consist of a list of API parameters encoded as key-value
pairs. The validation logic has been pulled into a new
ratechecker/validation.py module but has not been changed. As before,
the expected results are loaded from an XML file in the data archive.

This commit makes one slight change to how API parameters are validated.
The API always requires a `loan_amount` parameter which is the amount of
the loan being requested. It can take a `price` parameter which is the
cost of the house being purchased or refinanced; if this is provided,
then an `ltv` parameter is computed that represents the loan-to-value
ratio. The API validation scenarios, however, include this `ltv`
parameter directly, and don't include `price`. To make the API more
straightforward and consistent, this commit includes a minor change to
how parameters are validated that require exactly one of `price` or
`ltv`. New unit tests have been added to verify this change, and it
won't have any impact on how the API behaves on consumerfinance.gov.

This commit explicitly does not attempt to improve the overall load
process whereby data is copied to temp tables and then copied back on
failure. This work is scheduled for a future task and so was not
included here so as to limit scope of this particular change.