Owning a Home API annual countylimit update
At the end of every year, the owning-a-home API's county-limit data
needs to be updated. This PR adds data files and an updated fixture
that will load new limits.

After the OaH API repo is updated and re-imported into cfgov-refresh,
the database limit records can then be updated with this manage.py command:

```
load_county_limits --confirm=y
```

This PR also removes the nasty initial_data fixture that loaded with every
deploy, and removes a convoluted dependency on csvkit