Deprecate CHANGELOG in favor of Github releases
The changelog was a manual process that was redudant to the github
release log. It was becoming out of date because it's easy to forget to
update it.

Before replacing the changelog with a deprecation message, I did a
review to ensure that all information from the changelog also exists in
the github release log.