Make regions load last.
Then the view can return no results if there are no region objects yet.
This still allows a return with a partial region load, but region is the
smallest table, and the least likely to be caught in mid-load.
This should short-circuit a query when any of the larger tables are incomplete.