Add Python 3 support to this repository
This commit adds (almost all that is needed for) Python 3 support for
this project. The unit tests run via tox are now configured to test
against both Python 2.7 and Python 3.6. The wheel built from releases of
this project is now built as --universal.

Support for Python 2.6 has also been removed.

There's only one unit test left that fails, that relates to our use of
unicodecsv. Unfortunately this library doesn't seem to support Python
3.6, and appears to be abandoned:

https://github.com/jdunck/python-unicodecsv/pull/85#issuecomment-467981537