Fix annotated Decimals in Django 2.2
In Django 2.2, the default Decimal precision of `rate.base_rate` was NOT being maintained through string conversion in the `data` dictionary. This was the result of adding the `sum_of_adjvalue` annotation to the base_rate. This change is to use `quantize()` to round the annotated value to the same precision as the `base_rate`.