Improve error reporting in the data collection routine
Previously, if a problem crept into our base data -- such as the creation
of a new county FIPS code based on census changes -- the script would
simply report that some files were not created.

This change adds a check for a KeyError, and adds a note about what a five-digit
error code probably means. We've been tripped by a census change like this before,
and KeyError was the problem then as well.

The patch also changes handling of the year value to use a timezone-aware
datetime object, so that running the script on 12/31 (not naming any names here)
won't get tripped up by UTC time.